Update glossary.md
Added styling around defined terms and changed some definitions to better fit general usage.

In particular, I removed "who participates in the community" from the definition of *user*, as, while this may be encouraged or required in some use cases of the software, is not necessarily required in all. It's entirely plausible that a user, in a hypothetical use case, may be entirely passive and not participate in the community at all.